move existing gngeo configs to new location, and symlink. don't overwrite if there is an existing config

+++ scriptmodules/emulators/gngeopi.sh
@@ -29,14 +29,26 @@ function install_gngeopi() {
 function configure_gngeopi() {
     mkRomDir "neogeo"
 
-    # add default controls for keyboard p1/p2
-    mkdir -p "$home/.gngeo"
-    cat > "$home/.gngeo/gngeorc" <<\_EOF_
+    mkdir -p "$configdir/neogeo"
+
+    # move old config to new location
+    if [[ -d "$home/.gngeo" && ! -h "$home/.gngeo" ]]; then
+        mv "$home/.gngeo/"* "$configdir/neogeo"
+        rmdir "$home/.gngeo"
+    fi
+    
+    if [[ ! -f "$configdir/gngeo/gngeorc" ]]; then
+        # add default controls for keyboard p1/p2
+        cat > "$home/.gngeo/gngeorc" <<\_EOF_
 p1control A=K122,B=K120,C=K97,D=K115,START=K49,COIN=K51,UP=K273,DOWN=K274,LEFT=K276,RIGHT=K275,MENU=K27
 p2control A=K108,B=K59,C=K111,D=K112,START=K50,COIN=K52,UP=K264,DOWN=K261,LEFT=K260,RIGHT=K262,MENU=K27
 _EOF_
+    fi
+    
+    # symlink to new config location
+    ln -snf "$configdir/neogeo" "$home/.gngeo"
 
-    chown -R $user:$user "$home/.gngeo"
+    chown -R $user:$user "$configdir/neogeo"
 
     delSystem "$md_id" "neogeo-gngeopi"
     addSystem 0 "$md_id" "neogeo" "$md_inst/bin/gngeo -i $romdir/neogeo-gngeopi -B $md_inst/neogeobios %ROM%"
